Filter out untraceable functions on attaching
Curretly, ipftrace2 tries to attach eBPF programs to the functions it
found from debug information regardless if it is untraceable (inlined,
marked as "notrace", etc...) or not. As a result, users see that such
functions are counted as "failed" to attach which seems weired. Another
bad point of this behavior is it produces a lot of "trace_kprobe: Could
not probe notrace function ..." messages which is annoying.

To avoid such behavior, filter out the untraceable functions on
attaching. We can determine the function is "traceable" by looking
"/sys/kernel/debug/tracing/available_filter_functions". The untraceable
functions are displayed as "untraceable".