Abstract the function to get function address
In kprobe, we can get the address of the function we
are tracing by using PT_REGS_IP. However, this is not
the case for fentry/fexit which don't have access to
the pt_regs. Instead, we need to rely on the
bpf_get_func_ip helper to get it.

Thus, depending on which tracing method we are using,
we need to change the way to get the function address.