Fix unsuitable names in script convention
Changed emit function to program variable and dump function to decode
function. These functions are not matched to its functionality
(especially emit).

The emit function was originally defined as a function that generates
raw BPF bytecode. Over the time, it is changed to the function that just
returns static BPF ELF binary (as far as we know). Thus, it doesn't make
sense to be function anymore. Change it to "program" variable that
stores BPF ELF binary.

The dump is not so good naming, because what it does now is decoding the
trace event binary and returns structured Lua table. So, just name it as
"decode" should be suitable.

These are the breaking changes from user's point of view. However, it
shouldn't be difficult to migrate to that since the semantics of them
are not changed. It should only require simple renaming.