Add smoke test
Thanks to the readiness probe feature, now it is straightforward to
write tests. Add smoke test for a minimum disruption detection. There
are currently two environments for runnig tests. First one is a CentOS
Stream 8 environment which has kernel v4.18.0. This is our minimum
kernel requirement. Another one is a Arch Linux environment which is
shipped with state of the art kernel.

This commit also adds new GitHub Action workflow to run smoke test for
every push to the master as well as for PRs.

Signed-off-by: Yutaro Hayakawa <yhayakawa3720@gmail.com>