Make MAX_SKB_POS variable
Currently, MAX_SKB_POS is fixed to 5, because kprobe can only access to
the first 5 arguments of the function. However, ftrace supports 12
arguments, so it's worth to support 12 arguments if possible to maximize
the tracing coverage.

Also, as reported in #9, we have a bug that we skip tracing the function
which has more than MAX_SKB_POS arguments which doesn't make sense. What
we really need here is avoid tracing functions which has more than 12
arguments for ftrace backend only, because ftrace cannot trace such
functions.

To resolve this, we introduce 2 new parameters for symsdb
initialization, max_args and max_skb_pos. max_args is a maximum number
of function argument which will be set to 12 when we are using ftrace
backend. This parameter doesn't make sense for kprobe backends, so it
will be set to INT_MAX which is practically unlimited. max_skb_pos is a
maximum skb position in the function arguments which will be set to 12
for ftrace backend and 5 for kprobe backends.

Fixes: #9

Signed-off-by: Yutaro Hayakawa <yhayakawa3720@gmail.com>