Display kernel address and skb position in -l option
This also slightly changes how ipftrace2 displays the symbol name (it
starts to put extra spaces at the beginning). Also, it displays the
header, but header text is written into stderr and symbol informations
are written into stdout, so the users parsing the output with UNIX shell
tools shouldn't be affected.

Signed-off-by: Yutaro Hayakawa <yhayakawa3720@gmail.com>