Handle symbol name resolution error correctly
When the address => symbol name conversion fails, currently we report
error immediately and terminate the tracer. This was based on the
assumption that address => symbol name conversion should always succeed.

However, when somehow kernel's PT_REGS_IP or bpf_get_func_ip() helper
don't return the correct value which doesn't match to the address of
kallsyms, it is possible to fail.

To resolve this issue, we handle symbol name resolution error and report
the error by displaying "(unknown)" as a "symbol" like we did in the
past.

Signed-off-by: Yutaro Hayakawa <yhayakawa3720@gmail.com>