Introduce an experimental BPF-based extension module
Introduce an experimental BPF-based extension module. The purpose of
this feature is the same as Lua-based extension module, but users only
need to write BPF program. This reduces the context switch and the
burden of managing both of eBPF program and Lua script.

The format of the BPF program is almost the same as Lua-based extension.
The only difference is users needs to annotate the struct that
represents the format of the perf event payload with __ipft_event_struct
variable declaration like below.

```
struct gso_info {
  unsigned int len;
  uint16_t gso_size;
  uint16_t gso_segs;
  uint32_t gso_type;
} __ipft_event_struct __attribute__((section("ipft_skip")));
```

This annotation tells the ipftrace2 event decoder that this struct
represents the perf event format. So that it can know the struct format
from BTF and decode it.

We use global variable for now. Otherwise, clang doesn't keep the type
information (struct gso_info in this case) to the BTF. We put the
variable into the special ipft_skip section since we don't want to
instanciate it (libbpf doesn't interpret the unrecognized section). We
will hide this declaration into the macro in the subsequent commits.

The user space program creates the "printer" of the perf event from BTF
and generates the output by key-value pair like this.

```
... ipv4_conntrack_in ( len: 2034 gso_size: 1318 gso_segs: 2 gso_type: 1 )
```

The key name is took from the member name of the struct and the value is
interpreted from the member types (in this case, all of the members have
an integer BTF type, so it prints the integer value. Currently, we
support printing the pointer and signed/unsigned integer type with boolean
and char encoding. In the future, we'll support the annotation that
tells the format of the field (e.g. hex, ipv4 address, ipv6 address,
enum, enum flags, etc...).

Signed-off-by: Yutaro Hayakawa <yhayakawa3720@gmail.com>