Add a command to generate BPF module skeleton
Add a new command to generate BPF module skeleton. This skeleton
contains the short-hand macros to to write BPF module. This skeleton is
hard-coded into the code, so we can maintain it together with the actual
code to interpret the macros. When we add a new macro, we also update
the skeleton.

Signed-off-by: Yutaro Hayakawa <yhayakawa3720@gmail.com>