Implement compile and run for BPF extension
Implement easy short-hand which compiles the BPF extension C source
code and directly use the output for tracing. It requires the clang to
be present on the local system, but reduces the burden of separating the
compile and execution.

It internally uses popen(3), thus the C compiler will be executed from
shell. Users can override the C compiler command (default is clang) with
CC environment variable and can append extra flags using CFLAGS
environment variable.

Signed-off-by: Yutaro Hayakawa <yhayakawa3720@gmail.com>