Automatically generate/check cmdref and BPF extension skeleton on docs
Currently, we manually generate the cmdref in README.md and BPF extension
skeleton in docs/bpf_extension.md while these can be automatically generated
from ipft command. To avoid the manual generation, introduce a new make target
`docs` which fills the auto-generated part of the docs with envsubst.

If we can autogenerate the docs, now we can automatically detect out-of-sync
docs as well. Introduce a new make target `docs-check` which first generates
docs with `docs` and compare the diff with the latest commit. Ideally, this
should be added to the pre-commit hooks of the developers and auto should be
checked for every push with GHA.

Signed-off-by: Yutaro Hayakawa <yhayakawa3720@gmail.com>