Don't build ipft-builder for every push
I don't remember why it is built for every push. Probably, I wanted to
use the latest builder image in the smoke-test workflow, but it's broken
since they will run in parallel, so now it always uses the previous
version of the builder. Remove this broken logic and build the image for
every smoke-test execution.

Signed-off-by: Yutaro Hayakawa <yhayakawa3720@gmail.com>