Preserve file owner on build within container
When we compile source within the container, the owner of the output
files are overwritten with root:root. Same problem can happen in
formatting as well (in case of the formatting, it's even worse since it
changes the owner of the source file itself). Avoid these issues by
running docker command with -u option.

Signed-off-by: Yutaro Hayakawa <yhayakawa3720@gmail.com>