Support custom formatting in BPF extension
Add a new mechanism to support custom formatting of event struct in BPF
extension. Users can format the struct members by specifying the special
BTF decl tag like this.

```
struct event {
	uint32_t formatted __atrribute__((btf_decl_tag("ipft:fmt:hex")));
	uint32_t unformatted;
};
```

And this produces output like this.

```
... ipv4_conntrack_in ( formatted: 0x123, unformatted: 123 )
...   nf_conntrack_in ( formatted: 0x123, unformatted: 123 )
```

In this case, we specify "hex" as a format. "ipft:fmt" is a common
prefix. Formats are predefined by ipftrace2 and an only format supported
right now is "hex". In the future, we'll introduce more advanced
formatters.

Signed-off-by: Yutaro Hayakawa <yhayakawa3720@gmail.com>